add paypal payment functionality, add my order page

+++ payment/views.py
@@ -5,6 +5,9 @@ from cart.cart import CartService
 from cart.models import Cart, CartItem
 from django.http import JsonResponse
 
+from django.core.mail import send_mail
+from django.conf import settings
+
 # Create your views here.
 
 
@@ -13,28 +16,27 @@ def checkout(request):
     if request.user.is_authenticated:
         shipping = ShippingAddress.objects.filter(user=request.user).first()
         form = ShippingForm(instance=shipping)
-        if request.method == "POST":
-            form = ShippingForm(request.POST, instance=shipping)
-            if form.is_valid():
-                shipping_user = form.save(commit=False)
-                shipping_user.user = request.user
-                shipping_user.save()
-                return redirect("checkout")
 
         return render(request, "payment/checkout.html", {"form": form})
-
     else:
-        return render(request, "payment/checkout.html",{"form": form})
+        return render(request, "payment/checkout.html", {"form": form})
 
 
 def payment_success(request):
     if request.user.is_authenticated:
         user_cart = Cart.objects.filter(user=request.user).first()
+        if not user_cart:
+            return redirect("payment_failed")
         user_cart.delete()
 
     else:
-        guest_cart = Cart.objects.filter(session_key=request.session.session_key).first()
+        guest_cart = Cart.objects.filter(
+            session_key=request.session.session_key
+        ).first()
+        if not guest_cart:
+            return redirect("payment_failed")
         guest_cart.delete()
+
     return render(request, "payment/payment-success.html")
 
 
@@ -56,25 +58,72 @@ def complete_order(request):
             filter(None, [address1, address2, city, state, zipcode])
         )
 
+        print(shipping_address)
+
         cart = CartService(request)
 
         total_cost = cart.get_total()
+        product_list = []
 
         if request.user.is_authenticated:
-            order = Order.objects.create(full_name=name, email=email,shipping_address=shipping_address,amount_paid=total_cost, user=request.user)
+            order = Order.objects.create(
+                full_name=name,
+                email=email,
+                shipping_address=shipping_address,
+                amount_paid=total_cost,
+                user=request.user,
+            )
 
             for item in cart:
-                OrderItem.objects.create(order = order, product = item["product"], quantity=item["quantity"], price = item["total"], user = request.user)
+                OrderItem.objects.create(
+                    order=order,
+                    product=item["product"],
+                    quantity=item["quantity"],
+                    price=item["total"],
+                )
+                product_list.append(
+                    {"name": item["product"].title, "quantity": item["quantity"]}
+                )
 
         else:
-
-            order = Order.objects.create(full_name=name, email=email,shipping_address=shipping_address,amount_paid=total_cost)
-
+            order = Order.objects.create(
+                full_name=name,
+                email=email,
+                shipping_address=shipping_address,
+                amount_paid=total_cost,
+            )
             for item in cart:
-                OrderItem.objects.create(order = order, product = item["product"], quantity=item["quantity"], price = item["total"]) 
-        
+                OrderItem.objects.create(
+                    order=order,
+                    product=item["product"],
+                    quantity=item["quantity"],
+                    price=item["total"],
+                )
+                product_list.append(
+                    {"name": item["product"].title, "quantity": item["quantity"]}
+                )
+
+        body = (
+            "Hi!\n\n"
+            + "Thank you for placing your order\n\n"
+            + "Please see your order below:\n\n\n"
+            + "\n".join(
+                f"{product['name']} X {product['quantity']}\n"
+                for product in product_list
+            )
+            + f"\n\nTotal paid: ${cart.get_total()}"
+        )
+
+        send_mail(
+            "Order received",
+            body,
+            settings.EMAIL_HOST_USER,
+            [email],
+            fail_silently=False,
+        )
+
         order_success = True
 
-        response = JsonResponse({'success':order_success})
+        response = JsonResponse({"success": order_success})
 
         return response
